Add command import_zupc
It can be launch with manage.py import_zupc
It will download contours from OSM, import it in a temporary table.
Then it will read /tmp/zupc directory. This directory can have one
directory per ZUPC, in each of it reads the content of *.list files,
each line of these files has to be a INSEE code, it will join every
insee of these files.