Always save taxi phone number on PUT /hails/:id
When PUT /hails/:id is called with the status "accepted_by_taxi", the
taxi's phone number must be provided.

Before, it was only saved for X-Version == 2. Now, it is saved for any
version >=2.

I prefer not to enforce the parameter for any version (< 2), as it would
break the backward compatibility, and I don't know if all clients
provide the X-Version header.