Partially revert 6c9b736f4eb3aa
Our documentation states the taxi's phone number must be provided when
the taxi accepts the trip and the status changes to accepted_by_taxi.

Some clients do not provide it because they returned it before, when we
called their API to send the request.

Make the parameter mandatory only in the case we do not already have the
number.