Do not raise if ADS, Driver or Taxi have duplicates
Objects ADS, Driver and Taxi should have composite primary keys to
prevent duplicates. It is not the case in production, and we find for
example several ADS with the same numero and insee code.

Instead of fixing the database, we prefer to keep data as-is for now and
ignore "old" entries to keep backward compatibility.

In the future, we should:

- remove duplicates from ADS, Driver and Taxi
- add primary keys
- change this commit to update .order_by(id.desc()).first() back to .one_or_none()
- remove the new tests
- remove VehicleDescription duplicate entries, which only exist for
  entries created by neotaxi

To check duplicates, run:

sql> SELECT
     numero
     , insee
   FROM "ADS"
   GROUP BY
     numero
     , insee
   HAVING
     COUNT(*) > 1