Do not magically change hail status when taxi status changes
In the previous behavior:

- when a hail was accepted_by_taxi
- then accepetd_by_customer
- then the taxi status changes to "occupied"

we assumed the client was on board, and we automatically changed the
hail status to "customer_on_board" (even if we didn't set the new hail
status explicitly).

Also:

- when a hail was "customer_on_board"
- then taxi changes status to "free"

we assumed the hail was done, and we changed the hail's status to
"finished".

These magic changes are hard to debug and are actually useless. It is
now possible to have an inconsistent situations, for example the taxi can
be "free" but have a current hail with the status "customer_on_board",
but at least changes are explicit.