migrations: clean up hail endpoint columns
They were added in 2015, but only the production one was used in the
end.

The other ones are confusing the user, as both production and
development environment use the same code base reading the same
production column.

Also making hail_endpoint_production not nullable to avoid mistakes in
the future.