Allow to get taxi location from hail from the beginning
It is now possible for search providers to get the taxi's location as
soon as the hail is emitted from GET /hails/:id. Before, it was not
possible to get the taxi's location until the hail was accepted (it was
possible a longtime ago, but we changed this behavior a few months ago).

One of our partners requires the location to be available. We do not
consider it is sensitive, since the search provider can get the taxi's
location through GET /taxis anyway.