schemas: short path to restore backwards compatibility
I was writing the alternative explicit path, but it's much much more
verbose. This may be for the best to keep a reasonable amount of code.

I think we're safe on the Ref* schemas, where only required fields are
accepted to create a taxi, but we won't catch typos on status or radius
anymore, the only editable and optional fields. An acceptable trade-off.