stats: merging archive hails and the need for hail stats
ArchivedHail was a first attempt at collecting stats but for old hails.
When building stats for recent hails using timescaledb, I just happened
to reinvent ArchivedHail with a saner base.

Now we have a single source for querying stats, and hails can be deleted
at any time after their stats were collected.

This required the dev environment to use timescaledb-ha as a base image
(and thus moving to Ubuntu 22.04).