prepare for psycopg 3 but cancel at the last moment
It's more picky on types, which is both a blessing and a curse.

That makes it tedious with JSON and geometry types where everything has
to be cast. While there is some shapely support in psycopg, not a word
about integration from SQLAlchemy, and what about pool connections?

I guess psycopg 3 is just too young, and psycopg2 is still active and
popular.